data/spark/src/requirements.txt
pandas
numpy
yfinance
python-dotenv
pymongo
pyspark
openpyxl
finnhub-python